Set project version to 0.0.0
This way when the publish workflow is run it can increment the version
number to 0.1.0 for the first release.

Also adjust the comment in the publish workflow; letting the workflow
automatically select the next version number is probably not a good idea
because it can only detect API changes but not incompatible implementation
changes.