Dual license under MIT and Apache license
I am personally not entirely convinced by this and I think it would be
simpler to just license under the MIT license (also because the Apache
license text is quite verbose). However, for increased compatibility
with other Rust projects, choose this dual licensing because it seems
to be most prevalent.

I am however not planning on adding any license headers to the source
files at the moment.