Fix `JsonStreamReader::next_name` sometimes returning wrong results
Fixes #10

This fix delays consuming the `:` behind the name until the corresponding
member value is consumed.

Another solution would have been to add a method to `BytesValue` to
force it to use `value_bytes_buf`, copying the value from `buf` in case that
is currently used. And then call that method from `next_name`.