Improve string value reader and writer error behavior
Now when an error occurs, any subsequent read and write attempts will
return an error as well. This might still not fully satisfy the
requirements of `Read::read` and `Write::write`
(see https://users.rust-lang.org/t/error-handling-for-custom-read-read-write-write-implementation/101742)
but it is at least less error-prone than the previous behavior in case
users assume they can continue reading from a `Read` or writing to a
`Write` after an error occurred.