[simple-api] Rename reader and writer methods
For the reader the `next_...` methods might have been a bit confusing for
the cases where only a single value can be read, such as the top-level value.

For the writer not having a consistent prefix for the method names might have
made it more difficult to directly find all writing methods, e.g. during code
completion.

Could also consider doing similar renaming for the 'advanced API' in the future.