[simple-api] Reduce usage of `Box<dyn Error>` in public API
This will make it easier for users to inspect the errors.

Ideally would also replace usage of `Box<dyn Error>` for methods taking
a function, e.g. `write_array` or `read_array`, but that seems to make
usage rather cumbersome because usage of `?` is then not easily possible
anymore for custom error types in the functions.