Document seeking behavior for empty JSON paths
For `seek_to` the description is a bit more verbose because the method
always calls `peek`, so there is no guarantee that the reader stays at
the current position.

For now only extend the 'Advanced API' documentation but not the
'Simple API' one since an empty JSON path might be a rare corner case.